ktlo(common): split evm api and transaction into submodules
